Bam.Net.Core 2 Rc1

Culmination of a couple years of work, many bug fixes and functionality enhancements.

+++ _tools/bake/ConsoleActions/ConsoleActions.Zip.cs
@@ -30,7 +30,8 @@ namespace Bam.Net.Bake
                 }
             }
             DirectoryInfo dirInfo = new DirectoryInfo(recipe.OutputDirectory);
-            string fileName = $"{recipe.Name}.zip";
+            string defaultFileName = $"{recipe.Name}.zip";
+            string fileName = GetArgumentOrDefault("zip", defaultFileName);
             string output = $"./{fileName}";
             FileInfo outputFile = new FileInfo(output);
             if (Arguments.Contains("output"))
@@ -39,17 +40,22 @@ namespace Bam.Net.Bake
                 outputFile = new FileInfo(Path.Combine(output, "..", fileName));
             }
 
-            if (outputFile.Exists)
+            if (outputFile.Exists && outputFile.Name.Equals(defaultFileName))
             {
-                OutLineFormat("File {0} exists, deleting...", ConsoleColor.DarkYellow, outputFile.FullName);
+                Message.PrintLine("File {0} exists, deleting...", ConsoleColor.DarkYellow, outputFile.FullName);
                 Thread.Sleep(300);
                 File.Delete(outputFile.FullName);
                 Thread.Sleep(300);
             }
-            OutLineFormat("Zipping {0} to {1}...", ConsoleColor.DarkYellow, recipe.OutputDirectory, outputFile.FullName);
+
+            if (File.Exists(outputFile.FullName))
+            {
+                File.Move(outputFile.FullName, outputFile.GetNextFile().FullName);
+            }
+            Message.PrintLine("Zipping {0} to {1}...", ConsoleColor.DarkYellow, recipe.OutputDirectory, outputFile.FullName);
             ZipFile.CreateFromDirectory(new DirectoryInfo(recipe.OutputDirectory).FullName, outputFile.FullName);
-            OutLineFormat("\r\nZipped {0} to {1}", ConsoleColor.Green, recipe.OutputDirectory, outputFile.FullName);
+            Message.PrintLine("\r\nZipped {0} to {1}", ConsoleColor.Green, recipe.OutputDirectory, outputFile.FullName);
             Thread.Sleep(300);
-        } 
+        }
     }
 }
